thea: Disable secure discard
Some users have the habit of wiping the cache every now and then.
Because of the secure discard, this operation takes quite a lot, to
the point that some users may think recovery is just stuck and force
an emergency reboot while formatting the emmc. Disable secure wipes
to speed things up.

Change-Id: I217056b0457c94b8715abc7edd66b4f8bf42d472